редактирование роутов, валидация, ошибки

+++ controllers/movies.js
@@ -1,4 +1,7 @@
 const Movie = require('../models/movie');
+const { BadRequestError } = require('../errors/BadRequestError');
+const { AccessDeniedError } = require('../errors/AccessDeniedError');
+const { NotFoundError } = require('../errors/NotFoundError');
 const { STATUS_CODES } = require('../utils/constants');
 
 const getMovie = (req, res, next) => {
@@ -37,17 +40,31 @@ const addMovie = (req, res, next) => {
     owner: req.user._id,
   })
     .then((movie) => res.status(STATUS_CODES.CREATED).send(movie))
-    .catch((err) => next(err));
+    .catch((err) => {
+      if (err.name === 'ValidationError') {
+        return next(new BadRequestError('Переданы некорректные данные'));
+      }
+      return next(err);
+    });
 };
 
 const deleteMovie = (req, res, next) => {
   Movie.findById(req.params.movieId)
-    .then(() => {
+    .orFail(new NotFoundError('Фильм не найден'))
+    .then((movie) => {
+      if (movie.owner.toString() !== req.user._id) {
+        throw new AccessDeniedError('Нет прав для удаления');
+      }
       Movie.findByIdAndRemove(req.params.movieId)
-        .then(() => res.send({ message: 'Удалено' }))
+        .then(() => res.send({ message: 'Фильм удален' }))
         .catch(next);
     })
-    .catch((err) => next(err));
+    .catch((err) => {
+      if (err.name === 'CastError' || err.name === 'ValidationError') {
+        return next(new BadRequestError('Переданы некорректные данные'));
+      }
+      return next(err);
+    });
 };
 
 module.exports = {
